Class redesign and make rspec give useful errors
This does some basic class redesign to allow rspec to actually test a
bit better. It does not, however, fix all the rspec issues, nor the fact
that several classes are now being improperly called from init.pp ;)

It does, however, get the rspec failures to something that is easy to
start tracking down and should be a decent starting point on learning
rspec a bit better.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>